+++ src/AngleSharp.Js.Tests/Fixtures/Html5Test/whichbrowser-stub.js
@@ -1,0 +1,22 @@
+function WhichBrowser() {
+}
+
+WhichBrowser.prototype.isBrowser = function () {
+    return false;
+};
+
+WhichBrowser.prototype.isOs = function () {
+    return false;
+};
+
+WhichBrowser.prototype.isDevice = function () {
+    return false;
+};
+
+WhichBrowser.prototype.isType = function () {
+    return false;
+};
+
+WhichBrowser.prototype.isEngine = function () {
+    return false;
+};